Add a Thorough Guide for Beginners
This new file contains a guide that beginners can follow to install Heads without any prior experience. 

After struggling to install Heads, I noticed that the existing installation documentation is scattered. Some is contained in Building.md, Installing-Heads.md, Clean-the-ME-firmware.md and in other files. The files you need to reference if you are installing for the first time are not ordered and you have to move back and forth.

While it is good that the detailed documentation is there for those who can understand it, I think that there should be at least a succinct first installation guide for new users. The guide I have proposed is not complete and I plan to add pictures that further detail the steps. However, I think that even in its current state it would be helpful to the project.